requirements/requirements.txt
requests == 2.31.0 
pandas == 2.2.2
fastapi == 0.110.1 
uvicorn == 0.29.0
streamlit == 1.37.1
langchain == 0.2.14
langchain-chroma == 0.1.3
langgraph == 0.2.14
langchain-openai == 0.1.22
langchain-community == 0.2.12
python-dotenv
langfuse